etl/django_api/requirements.txt
django==4.2.5
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-split-settings==1.2.0
django-extensions==3.2.3

faker==19.6.2
flake8==3.9.0
flake8-bandit==2.1.2

psycopg2-binary==2.9.8
pytest==6.2.5
pre-commit==2.15.0
python-dotenv==0.17.0


bandit==1.7.2
gunicorn==21.2.0
uwsgi==2.0.22

elasticsearch==8.10.1
SQLAlchemy==2.0.9

requests==2.28.1
redis==4.3.4